Split JsonAPICondition into its schema and its condition. Include jsonApiConditionSchema in anyConditionSchema.
