+++ packages/taco/src/conditions/schemas/utils.ts
@@ -3,6 +3,7 @@ import { z } from 'zod';
 import { compoundConditionSchema } from '../compound-condition';
 
 import { contractConditionSchema } from './contract';
+import { jsonApiConditionSchema } from './json-api';
 import { rpcConditionSchema } from './rpc';
 import { sequentialConditionSchema } from './sequential';
 import { timeConditionSchema } from './time';
@@ -13,6 +14,7 @@ export const anyConditionSchema: z.ZodSchema = z.lazy(() =>
     timeConditionSchema,
     contractConditionSchema,
     compoundConditionSchema,
+    jsonApiConditionSchema,
     sequentialConditionSchema,
   ]),
 );
