+++ packages/taco/test/conditions/base/json.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest';
 
 import {
   JsonApiCondition,
-  JsonApiConditionSchema,
+  jsonApiConditionSchema,
 } from '../../../src/conditions/base/json-api';
 import { testJsonApiConditionObj } from '../../test-utils';
 
@@ -11,7 +11,7 @@ describe('JsonApiCondition', () => {
   describe('validation', () => {
     it('accepts a valid schema', () => {
       const result = JsonApiCondition.validate(
-        JsonApiConditionSchema,
+        jsonApiConditionSchema,
         testJsonApiConditionObj,
       );
 
@@ -26,7 +26,7 @@ describe('JsonApiCondition', () => {
       };
 
       const result = JsonApiCondition.validate(
-        JsonApiConditionSchema,
+        jsonApiConditionSchema,
         badJsonApiObj,
       );
 
@@ -43,7 +43,7 @@ describe('JsonApiCondition', () => {
       it('accepts conditions without query path', () => {
         const { query, ...noQueryObj } = testJsonApiConditionObj;
         const result = JsonApiCondition.validate(
-          JsonApiConditionSchema,
+          jsonApiConditionSchema,
           noQueryObj,
         );
 
@@ -54,7 +54,7 @@ describe('JsonApiCondition', () => {
       it('accepts conditions without parameters', () => {
         const { query, ...noParamsObj } = testJsonApiConditionObj;
         const result = JsonApiCondition.validate(
-          JsonApiConditionSchema,
+          jsonApiConditionSchema,
           noParamsObj,
         );
